update elastic search output
New entries will only be added to a single index. Indexes are prefixed
with "atmo" if the type is present. If the type contains the "atmo" then
the index will be based solely on the type field. When the type is empty
the entry is added to the uncategorized index. All other entries use the
default index.